Explicit cast to (unsigned int)
BN_num_bytes(bn) returns an int representing the size of a BIGNUM,
with no possibility of an error value [1].  As such, it is safe to
cast it to unsigned.

[1] https://www.openssl.org/docs/man1.0.1/crypto/BN_num_bytes.html

+++ lib/crypto/crypto_keys_subr.c
@@ -103,7 +103,7 @@ export_BN(BIGNUM * bn, uint8_t ** buf, size_t * buflen,
 	unsigned int bnlen;
 
 	/* Figure out how much space we need. */
-	bnlen = BN_num_bytes(bn);
+	bnlen = (unsigned int)BN_num_bytes(bn);
 
 	/* Add the required storage length to ${len}. */
 	if (*len + sizeof(uint32_t) < *len) {
